✅ fix MediaInfo tests; do not use 'the force'
To use the 'force' to ignore the absence of the physical file to read
the generate XML is not a good idea.

So i added the dummy files, and do not use 'the force'.

Background: former commit.
A mediainfo.xml should not be read on 'force' else we never ever would
read the file with mediainfo again...