🩹 [uds] disc folders in root are problematic, and should not be processed!
Within a datasource there are always more than one entities.
When we see a named disc folder (VIDEO_TS, BDMV), we automatically assume to be INSIDE a disc folder, and walk up one path.
So the MF path would be the datasource root.
This mixes everything up ;)
So we just skip such folders early...